dotmul GPU bug fixed

+++ Self-accelerating Processing Workflows/headers/DotMultiplicationModel.h
@@ -8,12 +8,12 @@ class DotMultiplicationModel : public ComputationalModel
 {
     public:
         int *localA, *localB;
-        long long *localC;
+        int *localC;
         int localL;
         DotMultiplicationModel();
-        DotMultiplicationModel(int *in1, int *in2, long long *out, int length);
+        DotMultiplicationModel(int *in1, int *in2, int *out, int length);
         virtual ~DotMultiplicationModel();
-        inline void setData(int *in1, int *in2, long long *out, int length){
+        inline void setData(int *in1, int *in2, int *out, int length){
             localA = in1; localB = in2; localC = out; localL = length;
             return;
         }
